WEB/SiteWebMeteo/api/requirements.txt
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dnspython==1.16.0
Flask==2.0.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.20.3
pandas==1.2.4
pycodestyle==2.7.0
pymongo==3.12.0
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
toml==0.10.2
urllib3==1.26.5
Werkzeug==2.0.1
xmltodict==0.12.0
